Relocate VecOrNone to common place
