+++ graphql/batch_mutations/src/lib.rs
@@ -2,16 +2,12 @@ mod batch_inbound_shipment;
 mod batch_outbound_shipment;
 mod batch_request_requisition;
 mod batch_stocktake;
-use self::batch_stocktake::*;
 use async_graphql::*;
 use batch_inbound_shipment::{
     batch_inbound_shipment, BatchInboundShipmentInput, BatchInboundShipmentResponse,
 };
 use batch_outbound_shipment::{
     batch_outbound_shipment, BatchOutboundShipmentInput, BatchOutboundShipmentResponse,
-};
-use batch_request_requisition::{
-    batch_request_requisition, BatchRequestRequisitionInput, BatchRequestRequisitionResponse,
 };
 
 #[derive(Default, Clone)]
@@ -41,17 +37,31 @@ impl BatchMutations {
         &self,
         ctx: &Context<'_>,
         store_id: String,
-        input: BatchRequestRequisitionInput,
-    ) -> Result<BatchRequestRequisitionResponse> {
-        batch_request_requisition(ctx, &store_id, input)
+        input: batch_request_requisition::BatchInput,
+    ) -> Result<batch_request_requisition::BatchResponse> {
+        batch_request_requisition::batch(ctx, &store_id, input)
     }
 
     async fn batch_stocktake(
         &self,
         ctx: &Context<'_>,
         store_id: String,
-        input: BatchStocktakeInput,
-    ) -> Result<BatchStocktakeResponse> {
-        batch_stocktake(ctx, &store_id, input)
+        input: batch_stocktake::BatchInput,
+    ) -> Result<batch_stocktake::BatchResponse> {
+        batch_stocktake::batch(ctx, &store_id, input)
+    }
+}
+
+pub trait VecOrNone<T> {
+    fn vec_or_none(self) -> Option<Vec<T>>;
+}
+
+impl<T> VecOrNone<T> for Vec<T> {
+    fn vec_or_none(self) -> Option<Vec<T>> {
+        if self.is_empty() {
+            None
+        } else {
+            Some(self)
+        }
     }
 }
